+++ src/lib/hono/routes/auth-routes.ts
@@ -1,0 +1,36 @@
+import {auth} from "@/lib/auth";
+import {APIError} from "better-auth/api";
+import {Hono} from "hono";
+import type {ContentfulStatusCode} from "hono/utils/http-status";
+
+const authRoutes = new Hono();
+
+authRoutes.get("/v1/sign-in", async (c) => {
+  try {
+    const response = await auth.api.signInTokens({
+      body: {
+        basicToken: c.req.header("Authorization") ?? "",
+      },
+    });
+
+    return c.json(response);
+  } catch (error: any) {
+    if (error instanceof APIError) {
+      return c.json(
+        {
+          error: error.body?.message,
+        },
+        error.statusCode as ContentfulStatusCode,
+      );
+    }
+
+    return c.json(
+      {
+        error: "Internal server error",
+      },
+      500,
+    );
+  }
+});
+
+export default authRoutes;
